chore: Move isDev to core directory
