+++ src/apps/main/background-processes/backups/BackukpWorker/BackupWorker.ts
@@ -1,7 +1,7 @@
 import { BrowserWindow } from 'electron';
 import Logger from 'electron-log';
 import path from 'path';
-import isDev from '../../../../utils/isDev/isDev';
+import isDev from '../../../../../core/isDev/isDev';
 
 export class BackupWorker {
   private static readonly DEV_PATH =
