+++ src/apps/main/windows/feedback.ts
@@ -2,7 +2,7 @@ import { BrowserWindow, ipcMain } from 'electron';
 
 import { preloadPath, resolveHtmlPath } from '../util';
 import { setUpCommonWindowHandlers } from '.';
-import isDev from '../../utils/isDev/isDev';
+import isDev from '../../../core/isDev/isDev';
 
 let feedbackWindow: BrowserWindow | null = null;
 export const getFeedbackWindow = () =>
