+++ src/apps/main/windows/migration.ts
@@ -2,7 +2,7 @@ import { BrowserWindow, ipcMain } from 'electron';
 
 import { preloadPath, resolveHtmlPath } from '../util';
 import { setUpCommonWindowHandlers } from '.';
-import isDev from '../../utils/isDev/isDev';
+import isDev from '../../../core/isDev/isDev';
 
 let migrationWindow: BrowserWindow | null = null;
 export const getMigrationWindow = () =>
