+++ src/apps/main/windows/widget.ts
@@ -5,7 +5,7 @@ import { TrayMenu } from '../tray/tray';
 import { preloadPath, resolveHtmlPath } from '../util';
 import { setUpCommonWindowHandlers } from '.';
 import { getIsLoggedIn } from '../auth/handlers';
-import isDev from '../../utils/isDev/isDev';
+import isDev from '../../../core/isDev/isDev';
 
 const widgetConfig: { width: number; height: number; placeUnderTray: boolean } =
   { width: 330, height: 392, placeUnderTray: true };
